Improving the logic.
If none of the desiredParams where passed in the URL the program would
get a runtime error “index out of range”

Now if the param is not there it will error specifying the param
correctly. It will now also return the param if it’s not required and
not blank. Because we loop through the desiredParams and not the params
passed we only return the requested ones and ignore the rest. Further
consideration is needed.